Move supervisor test to agentruntest package
The test does not belong in the probo package. Move it
alongside its shared helpers in pkg/agentruntest.
